Do not return value from Options::overrideDefault
The overrideDefault method should not be used to get the value of an
option - it does not make sense to do so because there is already the
withDefault method. So there is no need to return a value from the
method - removing the return reduces the number of lines in the method
and avoids potential confusion about where it should be used.

Also do not set value_used=true in overrideDefault, since it is possible
that the value might never be used (unlike in withDefault).